# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plant Doctor API",
        "description": "API for plant disease diagnosis and health tracking",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8005",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/diagnoses": {
            "get": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "List user diagnoses",
                "description": "GET /api/diagnoses\nList user diagnoses",
                "operationId": "6b5ce2fc3e77e59722fee40ad93b943b",
                "responses": {
                    "200": {
                        "description": "Diagnoses listed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Diagnosis"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Create a new diagnosis",
                "description": "POST /api/diagnoses\nCreate a new diagnosis",
                "operationId": "0383883ed7af41cccc6f1194efe6c5e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDiagnosisRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Diagnosis created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Diagnosis"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/diagnoses/{diagnosis}": {
            "get": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Get diagnosis details",
                "description": "GET /api/diagnoses/{diagnosis}\nGet diagnosis details",
                "operationId": "f4c60177da61eab28278d34669847b35",
                "parameters": [
                    {
                        "name": "diagnosis",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Diagnosis retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Diagnosis"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/diagnoses/{diagnosis}/request-expert-review": {
            "post": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Request expert review",
                "description": "POST /api/diagnoses/{diagnosis}/request-expert-review\nRequest expert review for a diagnosis",
                "operationId": "8aab01f3ec14b371f1dc775c3a36c32e",
                "parameters": [
                    {
                        "name": "diagnosis",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expert review requested",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Diagnosis"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Diagnosis": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "confidence_score": {
                        "type": "number",
                        "format": "float",
                        "example": 87.5
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "completed"
                    },
                    "ai_provider": {
                        "description": "AI service used for diagnosis",
                        "type": "string",
                        "example": "groq",
                        "nullable": true
                    },
                    "species": {
                        "properties": {
                            "name": {
                                "description": "Scientific name of the identified species",
                                "type": "string",
                                "example": "Rosa canina",
                                "nullable": true
                            },
                            "common_names": {
                                "description": "Common names of the species",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "dog rose",
                                    "escaramujo"
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "disease": {
                        "properties": {
                            "name": {
                                "description": "Common name of the detected disease",
                                "type": "string",
                                "example": "Black Spot",
                                "nullable": true
                            },
                            "scientific_name": {
                                "description": "Scientific name of the disease",
                                "type": "string",
                                "example": "Diplocarpon rosae",
                                "nullable": true
                            },
                            "severity": {
                                "type": "string",
                                "example": "medium",
                                "nullable": true,
                                "enum": [
                                    "low",
                                    "medium",
                                    "high"
                                ]
                            },
                            "symptoms": {
                                "description": "Observed symptoms",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Dark circular spots",
                                    "Yellowing leaves"
                                ]
                            },
                            "treatment": {
                                "description": "Recommended treatment",
                                "type": "string",
                                "nullable": true
                            },
                            "prevention": {
                                "description": "Prevention measures",
                                "type": "string",
                                "nullable": true
                            },
                            "catalog_match": {
                                "$ref": "#/components/schemas/Disease"
                            }
                        },
                        "type": "object"
                    },
                    "expert_verified": {
                        "type": "boolean",
                        "example": false
                    },
                    "expert_notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "plant": {
                        "$ref": "#/components/schemas/Plant"
                    },
                    "expert": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Disease": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Oídio"
                    },
                    "scientific_name": {
                        "type": "string",
                        "example": "Erysiphe cichoracearum",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "symptoms": {
                        "type": "string",
                        "nullable": true
                    },
                    "treatment": {
                        "type": "string",
                        "nullable": true
                    },
                    "prevention": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "category": {
                        "type": "string",
                        "example": "fungal",
                        "nullable": true
                    },
                    "severity": {
                        "type": "string",
                        "example": "medium"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "perenual_id": {
                        "description": "Perenual API ID",
                        "type": "integer",
                        "nullable": true
                    },
                    "last_synced_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error occurred"
                    }
                },
                "type": "object"
            },
            "Plant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mi Tomate"
                    },
                    "species": {
                        "type": "string",
                        "example": "Solanum lycopersicum",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Tomate del jardín",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "country_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "state_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "city_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "diagnoses_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "diagnoses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Diagnosis"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "StoreDiagnosisRequest": {
                "required": [
                    "plant_id",
                    "image"
                ],
                "properties": {
                    "plant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image": {
                        "type": "string",
                        "format": "binary"
                    },
                    "organ": {
                        "description": "Plant organ photographed for Pl@ntNet identification",
                        "type": "string",
                        "nullable": true,
                        "default": "leaf",
                        "enum": [
                            "leaf",
                            "flower",
                            "fruit",
                            "bark"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Hojas amarillas",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "user"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}